database/requirements.txt
# PostgreSQL Database Upload Script Requirements
# Install with: pip install -r requirements.txt

# PostgreSQL adapter for Python
psycopg2-binary>=2.9.0

# HTML to Markdown conversion (for the converter script)
html2text>=2020.1.16
beautifulsoup4>=4.9.0

# Optional: Progress bars and enhanced output
tqdm>=4.60.0
colorama>=0.4.4

# Optional: Configuration management
python-dotenv>=0.19.0
